+++ app/src/main/java/com/yapp/twix/di/FeatureModules.kt
@@ -1,9 +1,9 @@
 package com.yapp.twix.di
 
-import com.peto.onboarding.di.onBoardingModule
 import com.twix.home.di.homeModule
 import com.twix.login.di.loginModule
 import com.twix.main.di.mainModule
+import com.twix.onboarding.di.onBoardingModule
 import org.koin.core.module.Module
 
 val featureModules: List<Module> =
