+++ feature/onboarding/src/main/java/com/twix/onboarding/profile/component/OnBoardingTopbar.kt
@@ -1,4 +1,4 @@
-package com.peto.onboarding.profile.component
+package com.twix.onboarding.profile.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
